Refactor basic effects
