+++ src/local.js
@@ -1,17 +1,5 @@
-const _deps = Symbol.for('spect.deps')
+import { createEffect } from './util'
 
-export default function fx(fn, deps) {
-  // check [meta] deps
-  if (this[_deps]) if (!this[_deps](deps, () => {
-    destroy.forEach(
-      fn => fn && (fn.call && fn()) ||
-      fn.then && fn.then(fn => fn && fn.call && fn())
-    )
-  })) return this
-
-  let destroy = []
-
-  destroy.push(Promise.resolve().then(() => fn.call(this)))
-
-  return this
-}
+export default createEffect(function local(target) {
+  // TODO: serialize/parse any target
+})
